refactor(engine): fold variant-tautology OR rewrite into LowerStatement
  Move from `simplify/expr_or.rs` to a free function in `lower::expr_or`,
  invoked from an `Expr::Or` arm in `LowerStatement::visit_expr_mut` that
  runs the check pre-children.  Same pattern as #823 (eq-operand) and
  #835 (IN-list model→PK).

  The check pattern-matches on `Expr::IsVariant`; running pre-children
  fires the match before each `IsVariant` is lowered into a discriminant
  comparison.  Pulled out as a free function on `&Schema` so the unit
  tests can call it without scaffolding a `LowerStatement`; tests move
  from `simplify::tests::expr_or::variant_tautology` to
  `lower::tests::expr_or`.
