+++ crates/toasty/src/engine/simplify/expr_or.rs
@@ -1,5 +1,4 @@
 use super::Simplify;
-use bit_set::BitSet;
 use std::mem;
 use toasty_core::stmt;
 
@@ -55,11 +54,9 @@ impl Simplify<'_> {
             return Some(true.into());
         }
 
-        // Variant tautology: `is_variant(x, 0) or is_variant(x, 1)` covering all
-        // variants of the enum → `true`
-        if self.try_variant_tautology_or(expr) {
-            return Some(true.into());
-        }
+        // The variant-tautology rewrite (`is_variant(x, 0) or is_variant(x, 1)`
+        // covering all variants → `true`) fires in the pre-lowering
+        // `LowerStatement::visit_expr_mut` `Expr::Or` arm, not here.
 
         // OR-to-IN conversion, `a = 1 or a = 2 or a = 3` → `a in (1, 2, 3)`
         if let Some(in_list) = self.try_or_to_in_list(expr) {
@@ -174,50 +171,6 @@ impl Simplify<'_> {
         }
 
         false
-    }
-
-    /// Checks for variant tautology: when all variants of an enum are tested
-    /// via `IsVariant` on the same expression, the OR is always true.
-    ///
-    /// `is_variant(x, 0) or is_variant(x, 1)` over `{0, 1}` → `true`
-    fn try_variant_tautology_or(&self, expr: &stmt::ExprOr) -> bool {
-        // Find the first IsVariant to use as anchor
-        let Some(first) = expr.operands.iter().find_map(|op| match op {
-            stmt::Expr::IsVariant(iv) => Some(iv),
-            _ => None,
-        }) else {
-            return false;
-        };
-
-        let anchor_expr = &first.expr;
-        let model_id = first.variant.model;
-        let num_variants = self
-            .schema()
-            .app
-            .model(model_id)
-            .as_embedded_enum_unwrap()
-            .variants
-            .len();
-
-        let mut seen = BitSet::with_capacity(num_variants);
-
-        for operand in &expr.operands {
-            let stmt::Expr::IsVariant(iv) = operand else {
-                continue;
-            };
-
-            // Every `IsVariant` subject must be equivalent to the anchor:
-            // two syntactically different (or non-deterministic) subjects could
-            // disagree at runtime, so covering all variants of the anchor tells
-            // us nothing about them.
-            if !iv.expr.is_equivalent_to(anchor_expr) || iv.variant.model != model_id {
-                return false;
-            }
-
-            seen.insert(iv.variant.index);
-        }
-
-        seen.count() == num_variants
     }
 
     /// Converts disjunctive equality chains to IN lists.
