+++ crates/toasty/src/engine/simplify/tests/expr_or.rs
@@ -1,4 +1,4 @@
-use super::{test_schema, test_schema_with};
+use super::test_schema;
 use crate::engine::simplify::Simplify;
 use toasty_core::stmt::{Expr, ExprOr};
 
@@ -619,309 +619,5 @@ fn or_to_in_list_not_simplified_for_non_deterministic() {
     assert_eq!(expr.operands.len(), 2);
     for operand in &expr.operands {
         assert!(matches!(operand, Expr::BinaryOp(_)));
-    }
-}
-
-// Variant tautology tests
-
-mod variant_tautology {
-    use super::*;
-    use crate as toasty;
-    use crate::schema::Register;
-    use toasty_core::schema::app::VariantId;
-
-    #[derive(Debug, PartialEq, toasty::Embed)]
-    enum TwoVariant {
-        #[column(variant = 1)]
-        A,
-        #[column(variant = 2)]
-        B,
-    }
-
-    #[derive(Debug, PartialEq, toasty::Embed)]
-    enum ThreeVariant {
-        #[column(variant = 1)]
-        X,
-        #[column(variant = 2)]
-        Y,
-        #[column(variant = 3)]
-        Z,
-    }
-
-    #[test]
-    fn all_two_variants_becomes_true() {
-        let schema = test_schema_with(&[TwoVariant::schema()]);
-        let mut simplify = Simplify::new(&schema);
-
-        let model_id = TwoVariant::id();
-
-        // `is_variant(x, 0) or is_variant(x, 1)` → `true`
-        let mut expr = ExprOr {
-            operands: vec![
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 0,
-                    },
-                ),
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 1,
-                    },
-                ),
-            ],
-        };
-        let result = simplify.simplify_expr_or(&mut expr);
-
-        assert!(result.is_some());
-        assert!(result.unwrap().is_true());
-    }
-
-    #[test]
-    fn all_three_variants_becomes_true() {
-        let schema = test_schema_with(&[ThreeVariant::schema()]);
-        let mut simplify = Simplify::new(&schema);
-
-        let model_id = ThreeVariant::id();
-
-        // `is_variant(x, 0) or is_variant(x, 1) or is_variant(x, 2)` → `true`
-        let mut expr = ExprOr {
-            operands: vec![
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 0,
-                    },
-                ),
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 1,
-                    },
-                ),
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 2,
-                    },
-                ),
-            ],
-        };
-        let result = simplify.simplify_expr_or(&mut expr);
-
-        assert!(result.is_some());
-        assert!(result.unwrap().is_true());
-    }
-
-    #[test]
-    fn subset_of_variants_not_simplified() {
-        let schema = test_schema_with(&[ThreeVariant::schema()]);
-        let mut simplify = Simplify::new(&schema);
-
-        let model_id = ThreeVariant::id();
-
-        // `is_variant(x, 0) or is_variant(x, 1)` over 3-variant enum → no change
-        let mut expr = ExprOr {
-            operands: vec![
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 0,
-                    },
-                ),
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 1,
-                    },
-                ),
-            ],
-        };
-        let result = simplify.simplify_expr_or(&mut expr);
-
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn single_variant_of_two_not_simplified() {
-        let schema = test_schema_with(&[TwoVariant::schema()]);
-        let mut simplify = Simplify::new(&schema);
-
-        let model_id = TwoVariant::id();
-
-        // `is_variant(x, 0) or other` — only 1 of 2 variants → no tautology
-        let mut expr = ExprOr {
-            operands: vec![
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 0,
-                    },
-                ),
-                Expr::arg(1),
-            ],
-        };
-        let result = simplify.simplify_expr_or(&mut expr);
-
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn all_variants_with_extra_operands_becomes_true() {
-        let schema = test_schema_with(&[TwoVariant::schema()]);
-        let mut simplify = Simplify::new(&schema);
-
-        let model_id = TwoVariant::id();
-
-        // `is_variant(x, 0) or other_expr or is_variant(x, 1)` → `true`
-        let mut expr = ExprOr {
-            operands: vec![
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 0,
-                    },
-                ),
-                Expr::arg(5),
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 1,
-                    },
-                ),
-            ],
-        };
-        let result = simplify.simplify_expr_or(&mut expr);
-
-        assert!(result.is_some());
-        assert!(result.unwrap().is_true());
-    }
-
-    #[test]
-    fn different_inner_exprs_not_simplified() {
-        let schema = test_schema_with(&[TwoVariant::schema()]);
-        let mut simplify = Simplify::new(&schema);
-
-        let model_id = TwoVariant::id();
-
-        // `is_variant(x, 0) or is_variant(y, 1)` — different inner exprs → no change
-        let mut expr = ExprOr {
-            operands: vec![
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 0,
-                    },
-                ),
-                Expr::is_variant(
-                    Expr::arg(1),
-                    VariantId {
-                        model: model_id,
-                        index: 1,
-                    },
-                ),
-            ],
-        };
-        let result = simplify.simplify_expr_or(&mut expr);
-
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn duplicate_variants_not_simplified() {
-        let schema = test_schema_with(&[ThreeVariant::schema()]);
-        let mut simplify = Simplify::new(&schema);
-
-        let model_id = ThreeVariant::id();
-
-        // `is_variant(x, 0) or is_variant(x, 0) or is_variant(x, 1)` — duplicates
-        // cover only 2 of 3 variants after dedup → no tautology
-        let mut expr = ExprOr {
-            operands: vec![
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 0,
-                    },
-                ),
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 0,
-                    },
-                ),
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 1,
-                    },
-                ),
-            ],
-        };
-        let result = simplify.simplify_expr_or(&mut expr);
-
-        // Idempotent fires first, removing the duplicate. Then 2 of 3 → no tautology.
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn reversed_order_becomes_true() {
-        let schema = test_schema_with(&[TwoVariant::schema()]);
-        let mut simplify = Simplify::new(&schema);
-
-        let model_id = TwoVariant::id();
-
-        // `is_variant(x, 1) or is_variant(x, 0)` (reversed) → `true`
-        let mut expr = ExprOr {
-            operands: vec![
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 1,
-                    },
-                ),
-                Expr::is_variant(
-                    Expr::arg(0),
-                    VariantId {
-                        model: model_id,
-                        index: 0,
-                    },
-                ),
-            ],
-        };
-        let result = simplify.simplify_expr_or(&mut expr);
-
-        assert!(result.is_some());
-        assert!(result.unwrap().is_true());
-    }
-
-    #[test]
-    fn no_is_variant_operands_not_simplified() {
-        let schema = test_schema();
-        let mut simplify = Simplify::new(&schema);
-
-        // `arg(0) or arg(1)` — no IsVariant at all → no tautology
-        let mut expr = ExprOr {
-            operands: vec![Expr::arg(0), Expr::arg(1)],
-        };
-        let result = simplify.simplify_expr_or(&mut expr);
-
-        assert!(result.is_none());
     }
 }
